Implement stochastic effective volume exclusion

+++ src/consts.rs
@@ -1,3 +1,8 @@
+#[cfg(feature = "single")]
+use std::f32::consts::PI;
+#[cfg(not(feature = "single"))]
 use std::f64::consts::PI;
 
-pub const TWOPI: f64 = 2. * PI;
+use crate::Float;
+
+pub const TWOPI: Float = 2. * PI;
