+++ src/flowfield/mod.rs
@@ -1,6 +1,7 @@
+use crate::Float;
 use ndarray::{Array, Ix4};
 
-pub type FlowField3D = Array<f64, Ix4>;
+pub type FlowField3D = Array<Float, Ix4>;
 
 pub mod spectral_solver;
 pub mod stress;
